requirements.txt
requests>=2.28.0
google-api-python-client>=2.70.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
openai>=0.27.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
colorama>=0.4.6
tiktoken>=0.4.0
